Remove unused DateTime.Now call
The non-xplat code path (using CodeDom) was updated to not write
out a timestamp in order to get determinism. The resource string
no longer takes any arguments so the result of DateTime.Now does
not surface anywhere. This is just clean up that will prevent a
reader from wondering if there is still a determinism issue.