Update assembly loader
Related to issue #750.

We need the ability to load task-containing assemblies from arbitrary
locations on disk determined at run time, along with their dependencies.
It is assumed that their dependencies will be located next to them in
the file system, or in some well-known location.

Ideally we would handle the dependencies by hooking the `Resolving`
event on the default `AssemblyLoadContext`. However, due to
https://github.com/dotnet/coreclr/issues/5837 we can't currently do
this. Instead, we need to implement our own `AssemblyLoadContext`.

Indeed, this is what we currently do. However, we run into a problem when
a task assembly depends on, say, Microsoft.Build.dll (which is pretty much
guaranteed). Microsoft.Build will already be loaded into the default
`AssemblyLoadContext`, but since our task assembly was loaded into *our*
context, we are likely to be asked to load it again when the task needs
it. We currently handle this by looking for it in the app's base
directory, which seems like a good bet.

So we may end up loading this assembly twice, in two different
contexts--not great, but everything still works. Issue #750 comes into
play when Microsoft.Build.dll has been CrossGen'd to a native assembly.
Both contexts will find and try to load the native assembly--and this is
currently not allowed.

So to work around these other issues in assembly loading, we need to
maintain our own `AssemblyLoadContext` for the time being. However, we
now ask the default `AssemblyLoadContext` to first try satisfying the
load. If it succeeds, the load will occur in that context.

If it fails, we'll get an exception. Either the default context couldn't
locate an assembly with the given name, or located what *should* have
been the assembly but wasn't actually (e.g., a native file with the same
name, or a corrupted file) or may couldn't read the file due to
permission issues. If this occurs, we simply catch and swallow the
exception, and then attempt to find and load the assembly ourselves,
putting it in our context.

We could think about letting some of these through. For example, if the
default context located the assembly but couldn't read it
(BadImageFormatException or an exception representing a permission issue)
then it would probably be better to let that exception propagate rather
than catch or ignore it. It's not clear where we would draw the line,
though, and since this whole approach is a workaround for other runtime
issues I'd rather just keep it as simple as possible.