Sign product binaries
I discovered that our fully-signed builds were no longer actually
signing our product binaries. We use MicroBuild for strong-name and
Authenticode signing, and it looks for FilesToSign items in projects.
Searching through our code, I could only find these in a couple of
places related to resources assembly and VSIX production. These
correspond to the binaries that are, in fact, still being signed.

I'm not 100% sure what happened, but I think this is due to moving to a
new version of Microsoft.DotNet.BuildTools in commit db26c2d3. I assume
that it used to generate these items, and no longer does so. The fix
here is to explicitly add a FilesToSign item to each project producing
a product binary.