Update NerdBank.GitVersioning
Update NerdBank.GitVersioning from 1.4.30 to 1.5.46. We have a problem
where VSTS builds don't properly recognize that they're producing
builds for public release and continue to add the commit ID (for example
"-g12345678") to NuGet package names, which is undesirable. So far we
haven't been able to track down what's going wrong, and I'm hoping a
newer version of GitVersioning will solve the problem.

Even if it doesn't, upgrading to a newer version doesn't hurt.