Implement wild card import of project extensions.
These extensions will generally come from package management systems by writing to a file named:

  $(MSBuildProjectExtensionsPath)\$(MSBuildProjectFile).<SomethingUnique>.props

Each package management system should use a unique moniker to avoid collisions.  It is a wild-card import so the order of the import is based on the file system's ordering.

Fixes #784

- Included unit tests
- Fix tests that broke when BaseIntermedateOutputPath was moved.

Current import order is:

1. directory build props
2. project extension props
3. project extension targets
4. directory build targets

PORT TO master:

 Conflicts:
	src/XMakeTasks/Microsoft.Common.props
	src/XMakeTasks/UnitTests/Microsoft.Build.Tasks.UnitTests.csproj
	src/XMakeTasks/Microsoft.Common.CurrentVersion.targets