Beef up MockLogger's xunit-output tracking
MockLogger's default behavior is to emit the log to stdout, but that makes
debugging within xunit tricky, because stdout isn't tracked.
e6006f0 added some use (optional) uses of xunit's ITestOutputHelper, but
didn't direct the normal logging statements there. This change logs both
to the internal StringBuilder (used to for log asserts) and to
_testOutputHelper, if it's defined.

There is probably a perf hit here, because string.Format will eagerly
create a full string, where StringBuilder.AppendFormat does "the right
thing" and adds segments without allocating. But it seems worthwhile for
the increased diagnosability.