Remove #if RETAIL (always use that code)
When we moved to GitHub, we never set up the projects to define RETAIL for
release builds, as was done in the old repo. This led to a memory-use
regression because we were calling the debug method instead of the
optimized one.

Instead of defining RETAIL, we decided to just always use the optimized
code, to minimize differences between our release and debug builds.

Fixes #1001