Import all search paths when importing a wildcard (#944)
* Import all search paths when importing a wildcard
Changed the import search path logic to import matches from ALL folders
when a wildcard is specified. For example:
  - <Project Import="$(propsdir)\*.props">
    - (Where $(propsdir) has fallback to c:\1, c:\2)
  - MSBuild will import c:\1\*.props AND c:\2\*.props.

The motivation behind this change is when packages install to common
extension points in VS "15". Some packages may still install to the old
(global) MSBuildExtensionsPath location and others may install to the
app-local path. This change will allow for MSBuild to import the union of
those extension points.

Note: We discussed as a team having MSBuild "de-dupe" imports (i.e. if it
imported foo.props in the original location it would not import a fallback
foo.props). This method becomes very tricky with all the import options
(**\*.props, for example).