Update NuGet packages to include satellite assemblies
Update our NuGet packages to include the satellite assemblies for all
supported languages. This only covers assemblies that we are currently
included in NuGet packacges for which we also produce satellite
assemblies. Notably, we ship Microsoft.Build.Engine.dll and
Microsoft.Build.Conversion.Core.dll in NuGet packages but we do not yet
have localization setup for those.