Set AutoUnifyAssemblyReferences only if it hasn't already been set (#1019)
For .NET Core projects, we want to pass AutoUnify=true to the RAR task. That gets set through the AutoUnifyAssemblyReferences property which derives its value from the AutoGenerateBindingRedirects property. It’s odd to have AutoGenerateBindingRedirects=true in the NETCore.props given binding redirects don’t mean anything in .NETCore. This change makes it so that we can set AutoUnifyAssemblyReferences in the props file.