Make FileMatcher more friendly with forward and backward slashes (#1035)
* FileMatcher refactorings

* Replace properties with characters

On OSX, both of those properties return forward slash

* Update item tests to work with slashes

* FilematcherTests permute through different slashes

* Add Path equality to FileMatcher

To support includes and excludes with slashes and back-slashes

* Fix tests for crossplat IO

- update some tests and test utilities to use Path.Combine
- update test utilities to normalize slashes on assertions
- update GetItemsFromFragmentWithGlobs to reuse test infrastructure

update tests

* Tests for slash preservation

The slashes in the fixed directory part of a filespec containing a wildcard should be preserved. At least that's what the code does. Added unit tests to assert this behaviour

* Remove slash literals

* Replace Path.DirSep and Path.AltDirSep with literals