Don't crash in NeedSeparateAppDomain when no file
To go cross-platform, we moved from XmlTextReader to XmlReader, since only
the latter is supported in .NET Core. This introduced a subtle difference
in behavior: XmlTextReader doesn't throw on construction when given a
nonexistent file (it throws on the first .Read()), but XmlReader.Create
does throw immediately. That meant that the old error handling for
nonexistent file (unconditionally call `reader.Close()` in the `finally`
block) was no longer valid, because `reader` might be null--because the
statement where it gets assigned was what threw.

The solution is simple: only `Close` the reader if it's not null.