Refactor _CopySourceItemsToPublishDirectory target into ComputeFilesToPublish and CopyFilesToPublishDirectory
We should have a single place where consumers can extend publish to add and remove files to be published.  With this change, all files to be copied are computed during `ComputeFilesToPublish` and then they are copied to the publish directory in `CopyFilesToPublishDirectory`.  Users can put a target between those two targets to add/update/remove files from @(ResolvedFilesToPublish).

Fix #170