Update _dependencyPaths to be a HashSet
The `CoreClrAssemblyLoader._dependencyPaths` field represents all the directories that we should look in when trying to locate a task assembly's dependencies. Currently we store it as a `List<string>` which allows for the possibility that the same directory could be added multiple times, possibly with different casing.

Instead, we should use a `HashSet<string>` with a case-ignoring comparer. This ensures the same path will not end up in the collection multiple times. This better captures the intent of the collection.