Disable AddEntryStrongReference only for Mono
I do not understand why, but the presence of the early-return-if-on-Mono
check consistently causes a strong reference to `projectRootElement` to be
retained, even though it doesn't need to be. The generated MSIL from
master and xplat was identical except for an additional bool (local for
return of `IsMono()`) and the branch-to-early-return. That *shouldn't*
impact whether the object pointed to by `projectRootElement` is eligible
for garbage collection, but it appears to have had that effect somehow.

Disabling the test for Mono.

Enabling the test for .NET Core required creating a valid path for
Unix-like operating systems as well as Windows.

This is a workaround for #282.