Use property-format string in Import search paths (#1095)
Fixes #1056 by using `ProjectImportPathMatch`'s already-precomputed string
field that represents the string to search for in an Import. Computing it
every time through `GetProjectImportSearchPaths` was wasteful.

This required a number of changes:
 * Converted `ProjectImportPathMatch` to a class so it could be passed by
   reference, so it could be
 * Passed by reference to a translator so it could be
 * Made `INodePacketTranslatable` so it could be
 * Used in the toolset directly.

Since I removed the only use of `TranslateDictionaryList`, I deleted it
from both the interface and implementation of `NodePacketTranslator`.