Build the correct legacy packages for the correct framework
This is for the legacy packages only.

Tom had already excluded Microsoft.Build.Engine and Microsoft.Build.Conversion.Core from building packages for NetCore.  This excludes the other packages from building on net46 because they only package files into netstandard1.3 directories anyway.

This will make it possible for me to update the build definition to be publishing the legacy and new packages during the same build.