Introduce CrossTargeting extensibility imports
An "outer" build that imports `CrossTargeting.targets` can't use
`Microsoft.Common.targets`'s extensibility hooks, because some targets in
those folders may expect a fully-specified ("inner") build environment.
But NuGet (at least) needs to be able to extend the outer build--so that
it can restore to get to the inner builds.

This introduces a new folder
`Microsoft.Common.CrossTargeting.targets\ImportAfter` and a new property
`CustomAfterMicrosoftCommonCrossTargetingTargets` (and their Before
equivalents) that allow the CrossTargeting (outer) build to be extended.

It does not remove the old import mechanism, to give some time where both
approaches work so we can migrate.

Fix for #1061, but doesn't close it because we should remove the old
mechanism.