Update the runtime so all our tests will pass
In order to run tests we copy a whole bunch of stuff, including the core
libraries and runtime, to the test deployment folder. The specific versions
used are controlled by the project.json unpdated here. When our tests
execute msbuild, they do so by calling `<root>\Tools\dotnetcli\dotnet.exe <deployment folder>msbuild.exe`.
While there is a copy of the CoreCLR runtime under `<root>\Tools\dotnetcli`,
dotnet.exe actually ends up loading whatever version is in the deployment
folder. At the moment, that copy has a bug that prevents the runtime from
properly loading assemblies returned by handlers hooking the
`AssemblyLoadContext.Resolving` event, which is in turn blocking one of my
other bug fixes (for issue #750).

Ultimately it might be nice to have just one copy of the runtime that is used
for building and testing, but until we can get that sorted out we can avoid
the problem by bumping up the version of the runtime package we use.