Only one project updates the shared xlf files
This solves a race in the SyncXlf task. Because localization is per project, and because MSBuild uses a Shared directory with shared resources (which is not a project but gets linked in by all the other projects), each project will try to sync the shared xlf files and crash with concurrent writes.