Ignore whitespace text nodes when parsing projects (#1187)
This just ignores purely whitespace text nodes. The XML reader only looks at the first 4k of text to determine if it should be ignored. Our tests are generating 70k of whitespace so the XML reader gives up after 4k and assumes its a text node.

Closes #270

Fixes test failures in #1004 when this change is ported to master.