Re-design "type: platform" for MSBuild.
Instead of relying on a library being marked as "type: platform", we now have an MSBuild property that gives the name of the platform library.

By default, .NET Core projects use 'Microsoft.NETCore.App' for the platform library, but users or other nupkgs can set the property, if necessary.

Fix #17