Generate targets in metaproj
This allows users to access any target when building a solution.  The target may or may not exist but this will create the corresponding targets in the `.metaproj`.

I had to pass along the target names from the build request all the way down to the solution generator via ReadOnlyCollections.

Closes #1275