Remove _CheckForInvalidConfigurationAndPlatform InitialTarget (#1280)
* Only have _CheckForInvalidConfigurationAndPlatform target run before any build related target

This target is running before targets like Restore which don't need the check.  This change removes _CheckForInvalidConfigurationAndPlatform from InitialTargets and instead injects it before any of the BuildDependsOn, RebuildDependsOn, or CleanDependsOn targets.  This will work except when a user specifies a BeforeTargets of the first target in those lists.  To get around that, the user must specify a DependsOn of _CheckForInvalidConfigurationAndPlatform if they want the check to run before their target.

* Reword comment to hopefully make it more understandable

* Add Build, Rebuild, and Clean to the list of targets