Switch ProjectCollection.Version to using the assembly file version and not the product version
This is because, on FF, the ProductBuildPart seems to be returning the bit from the assembly file version and not the assembly version, which lead to ProjectCollection reporting a mix between the assembly version and the assembly file version.

As a consequence, for NetCore MSBuild, which does not use git versioning, we have to also add a static assembly file verison in assemblyinfo.cs. Otherwise the compiler default is set which is 0.0.0.0