Make SerializedException robust to load errors
`GetTypeInfo().Assembly` seems to produce bogus results at least some of
the time (noticed especially for XmlException). Since exceptions are
serialized by including the exception's type and assembly, and then
deserialized by reflecting over that assembly to find the type's
constructor, this could cause an unhandled exception if the assembly
couldn't be loaded.

This change wraps the attempt to find the constructor robust to
file-not-found errors--if the assembly can't be loaded, simply use the
preexisting FormattedError fallback case.

Fixes #1246.