Only use RtlMoveMemory on Full Framework
RtlMoveMemory is not available on Nano Server, so the runtime check
introduced in 60a585c isn't quite right: Nano Server is Windows but
doesn't have this API exposed.

Introduce a new feature flag for RtlMoveMemory, used only when building
for Full Framework. For Mono and .NET Core builds, use the simpler (and I
guess slower? Would like to see a benchmark)
Environment.GetEnvironmentVariables codepath.

Related to #1233.