Implement /WarnAsError to treat specified warnings as errors
Specify just /WarnAsError to have all warnings logged as errors as well as have the build fail.
Specify a list of warning codes to have just that set of warnings treated as errors as well as have the build fail.

Targets will still show as succeeded and the tasks will continue to execute but the overall build result will be a failure.

Related to #68 and will close in my next change to add /NoWarn.

* Feature switch FEATURE_RESOURCEMANAGER_GETRESOURCESET
* Support for command-line arguments having empty values