Error when importing from non-existent SDK
The if-exists checks meant that if you had a badly defined MSBuildSDKsPath
or an SDK name that didn't exist, you would get no error and just silently
get no imports. That's no good.

This adds the implicit import unconditionally, requiring all SDKs to have
both an Sdk.props and an Sdk.targets for the moment.