Do not reveal implicit elements through RawXml
The initial implementation caused errors when editing a project in Visual
Studio. The Roslyn Project System uses Project.Xml.RawXml to decide what
to show to the user, and handles edits by saving that file. When RawXml
had _Implicit elements, that made the first save create an invalid project
with a literal _Implicit attribute in its imports (as well as imports from
both the Sdk attribute and the now-present Import elements).

To fix this, do the same transformation for RawXml that was being done for
Save.