Consider Implicit status to find insertion loc
Fixes #1431.

If a project had an Sdk attribute (and thus implicit Import elements), calling DeepClone on it could throw:

```
Unhandled Exception: System.ArgumentException: The reference node is not a child of this node.
   at System.Xml.XmlNode.InsertAfter(XmlNode newChild, XmlNode refChild)
   at Microsoft.Build.Construction.ProjectElementContainer.AddToXml(ProjectElement child)
   at Microsoft.Build.Construction.ProjectElementContainer.InsertAfterChild(ProjectElement child, ProjectElement reference)
   at Microsoft.Build.Construction.ProjectElementContainer.DeepClone(ProjectRootElement factory, ProjectElementContainer parent)
   at Microsoft.Build.Construction.ProjectElementContainer.DeepClone(ProjectRootElement factory, ProjectElementContainer parent)
   at Microsoft.Build.Construction.ProjectRootElement.DeepClone()
```

This was because we traverse the ProjectElement tree to find a sibling XML element to insert next to. If an adjacent ProjectElement's XML element has no physical representation, it is unparented and therefore not a good target for manipulations.

Add an implicitness check to the existing is-valid-for-referencing check to cause the next-XML-tree-sibling search to pass over implicit elements.