[tests] Fix CommandLine tests on mono/osx
Add USE_MSBUILD_DLL_EXTN when .dll extension is being used, which is on
CoreCLR and Mono, currently. Earlier, FEATURE_RUN_EXE_IN_TESTS was being
overloaded with this meaning and the dll/exe indication was just a side
effect. FEATURE_RUN_EXE_IN_TESTS is meant to indicate whether a runtime
executable like mono/corerun should be used to run msbuild, in a test.

The tests used FEATURE_RUN_EXE_IN_TESTS to choose .dll/.exe extension
for msbuild, which meant that it would need to be enabled for Mono, but
we don't need to use mono binary explictly to run assemblies.

So, the USE_MSBUILD_DLL_EXTN was created to have the specific meaning.

Fixes:

      Microsoft.Build.UnitTests.XMakeAppTests.GetCommandLineQuotedExeOnPath [FAIL]
      Microsoft.Build.UnitTests.XMakeAppTests.GetCommandLine [FAIL]
      Microsoft.Build.UnitTests.XMakeAppTests.GetCommandLineQuotedExe [FAIL]
      Microsoft.Build.UnitTests.XMakeAppTests.ResponseFileInProjectDirectoryWinsOverMainMSBuildRsp [FAIL]
      Microsoft.Build.UnitTests.XMakeAppTests.ProjectDirectoryIsMSBuildExeDirectory [FAIL]