Ensure BaseIntermediateOutputPath has a trailing slash when set by user
We used to ensure the trailing slash in common targets but I moved this to common props.  This however broke the scenario when a user sets the property since the trailing slash is not ensured and the _CheckForInvalidConfigurationAndPlatform target would then fail.

I also introduced a unit test to ensure this doesn't regress again.

Closes #1509