Drop + reload changed projects in worker nodes
This is related to
https://devdiv.visualstudio.com/DevDiv/_workitems?id=366048 and
https://devdiv.visualstudio.com/DevDiv/_workitems?id=366077, and
possibly #1526.

In those bugs, a worker node is started and given build requests for the
same project while that project changes on disk. The first build works
(fresh load in the worker node). The second build works (content is
reloaded). But the third one fails--it throws because the project "has
unsaved changes" when it's trying to reload it again. That failure isn't
well communicated back to the main node, so this tends to not log
anything and just abruptly fail the build.

To fix this, I'm reverting the part of #1507 that hit in the worker
node--the only place where `_autoReloadFromDisk` is set to true in the
current world. Instead of using the new `Reload`, we just drop the
`ProjectRootElement from` the cache and load it from scratch, as was
done before.