Ensure built-in and imported targets are respected by the solution pr… (#1590)
* Ensure built-in and imported targets are respected by the solution project generator.

The goal is to take whatever target the user specified from the command-line and add it to the metaproj.  However, we add targets and targets can be imported.

This change ensures that command-line targets are added last after everything else and they're only added if they don't already exist.  This removes the need to know what built-in targets are there.  If the user specifies /t:Clean, then it's still added as a built-in target and at the end when adding user specified targets it is ignored.

Another case is for projects.  We add a targets like "Project_Name", "Project_Name:Clean" where the project name becomes a target.  However we also add sub-targets which need to be respected.  These are also now added before the user specified targets are added.  If you specify /t:Project_Name:Clean, it won't be added to the project as a user specified target.

Unit test has been updated to hopefully test all built-in targets.

Closes #1587