Allow GetProcessStartInfo to be overridden
Current ToolTask implementers are able to set values for previously
undefined environment variables and to override previously defined
variable values via the EnvironmentVariables property. They are not,
however, able to remove Environment Variables which should not be passed
down to the ToolTask's child process.

This inability to manage a ToolTask's environment creates issues
particularly in the builds of programs whose behavior is affected by the
process environment like https://github.com/dotnet/cli. A ToolTask which
invokes such a program needs to be able to selectively provide a `clean`
process, ensuring deterministic build behavior. Of course, sometimes it
makes sense for such variables to flow, but there is not a current
option for preventing that flow without re-implementing ToolTask.