Add checked-in build definitions
These build definitions will be used by pipebuild to build the product.

The signing now uses MicroBuild because the builds are being moved to
the DevDiv VSTS instance, and the Microsoft Code Signing system which
was used in MSEng is not supported there.

MicroBuild does not support the .NET Core MSBuild, so the full-framework
MSBuild is used to build the signing targets.  For this reason, the
signing logic is implemented as a standalone MSBuild project.