[mono] Use GetCommandLineArgs, also, to check for test runners
When running a test runner under mono, the process name is not helpful
to determine if we are running tests. Getting command line arguments,
OTOH, do help with this.

So, we bring back the old `GetProcessFromCommandLine()` method. And
also, add this as a parameter to the `ResetInstance_ForUnitTestsOnly(..)`
method.

This fixes the related tests on mono/osx.

FIXME: need to check if some of the callers to
`ResetInstance_ForUnitTestsOnly` need to use the new parameter, as the
comments suggest that they depend on the command line args.