Fix issue where ToolTask will find directories instead of files (#1532)
* Fix issue where ToolTask will find directories instead of files

The native function SearchPath will return a directory which isn't what we want.  The bug was only found on Windows because our FindOnPath() method was doing different things on Windows vs Non-windows.

This change unifies the implementation of FindOnPath() and only looks at files.  The method is only called by ToolTask since I removed the other reference to it so I moved it and the unit tests out of NativeMethodsShared and into ToolTask.

I've also written a test which failed before the fix and passes now.

Closes #1215