Adds a switch to skip emitting tools packages
This allows the tools to be skipped in the LZMA archive generation, which is important as the .tools folder's project.assets.json files include absolute paths to the locations of the cache locations that were avaiable at the time of restore. Since the LZMA archive creation restores to a folder and compresses the output, the project.assets.json files pointing to locations on the local machine or CI server would be included in the archive which would presumably break consumers.