Use dotnet-install from cli's master branch
This verison of dotnet-install supports more operating systems and lets
us install a working dotnet cli that supports building sdk on more
operating systems and versions.

As an example, currently this gets us a working cli on RHEL 7.3, which
isn't possible with the dotnet-install script on the cli's rel/1.0.0
branch.