Compare item names case-insensitively in lazy eval
Item and property names should always be compared using
`MSBuildNameIgnoreCaseComparer`, since they have always been case
insensitive.

When #868 introduced lazy evaluation, it added some dictionary-like data
structures that store item names, but used the default `string`
comparer, which is case sensitive. This resulted in a break to projects
that referred to an item with a different case from its original
definiton.

I attempted to scan for other instances of this pattern by searching for
Dictionaries of string using the default comparator in non-test
projects, and this looked like the extent of it.

Fixes #1751.