Make update-dependencies idempotent in the only update case
Today, if you run update-dependencies --Update and there are no
changes that need to be made (for example, you ran it previously) and
there are pending changes (i.e. git status shows diffs) the program
crashes with an unhandled exception because some code in VersionTools
is upset that there are pending changes since it thinks there should
be none.

Re-structure the code such that we don't call the method that's going
to throw when we are just updating dependencies. This also means
devs (or the composed build) who want to run this more than once do
longer have to reset changes beforehand.