Reduce nuspec dependencies for full framework(#1762)
We initially setup the nuspec reference generator to also update references for full framework, not just .net core (#1682). However, this brought in too many dependencies, some of them incorrect. This lead to restore issues: #1725

To fix this we're moving back to hand edited nuspec references for full framework.

Squashed commits:
* Revert "Add net46 dependencies to msbuild nuspecs (#1682)"

This reverts commit 08753ec2a7f596e38b175efa27f31a89bd3c8e1f.

* Add net46 dependencies to msbuild nuspecs

This enables users to publish their msbuild depending app for net46 and get the closure of dependencies needed for msbuild.
Resolves #1681